Migrate analyzer into analyze.standalone
