+++ talondoc/autogen/__init__.py
@@ -7,7 +7,7 @@ from typing import Optional, Union
 import jinja2
 import jinja2.sandbox
 
-from ..analyze import analyse_package
+from ..analyze.standalone import analyse_package
 from ..registry import Registry
 from ..registry.entries.user import UserPythonFileEntry, UserTalonFileEntry
 from ..util.progress_bar import ProgressBar
